house-keeping
- folder name changed from `fake-data-generator` to industrial standard naming convention of `synthetic-data-generator`. Both industry and academia popularly use the term `synthetic` over `fake`.
